requirements.txt
# Runtime
psycopg2-binary==2.9.11  # PostgreSQL adapter
python-dotenv==1.2.2     # Load .env into environment
flask==3.1.3             # Web framework
jinja2==3.1.6            # HTML templating (Flask dependency, pinned for stability)

# Dev / quality
black==26.3.1            # Code formatter (line length 120)
ruff==0.15.8             # Linter
pytest==9.0.2            # Test runner
pytest-mock==3.15.1      # Mock/patch helpers for pytest
